Completed FindNearestStation feature Minior change the type of attributes in database Minor change on class layout Mockup a few data into the database for testing the feature purpose

+++ FIT5120 - VicEnerG/Models/StationData.cs
@@ -15,7 +15,7 @@ namespace FIT5120___VicEnerG.Models
     public partial class StationData
     {
         public int Id { get; set; }
-        public short stationNumber { get; set; }
+        public int stationNumber { get; set; }
         public double monthRadiation { get; set; }
         public string month { get; set; }
